Rename alloc::arc to alloc::sync, to match std::sync
